♻️  Refactor: 차트 생성 코드 분리
